ITS: Fix unused argument warning for read_size
Fix unused argument warning for read_size.

The void cast tricks the compiler into thinking the argument is used.

I don't know what is wrong with the ifdef expression and thereby
causing the argument to be unread. But I don't see the need to
maintain these complex ifdef expressions when we can just cast it to
void for all platforms with no cost to the final binary.

Change-Id: I0a61834c852862e79fbf4654995bb08ac4ebb757
Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>